Conway Blocks Broke the undoable actions on add/remove block Refactor so no errors with Wall Wextra pedantic

+++ view/Font.h
@@ -21,7 +21,7 @@ namespace view {
         [[nodiscard]] TTF_Font* font() const;
 
         /****** CONST FUNCTIONS  ******/
-        [[nodiscard]] int widthOfString(const std::string& str) const;
+        [[nodiscard]] size_t widthOfString(const std::string& str) const;
 
       private:
         /****** DATA MEMBERS  ******/
